+++ snovault/elasticsearch/indexer.py
@@ -140,14 +140,14 @@ def index(context, request):
 
         if record:
             try:
-                es.index(index=namespaced_index, doc_type='indexing', body=indexing_record, id=index_start_str)
-                es.index(index=namespaced_index, doc_type='indexing', body=indexing_record, id='latest_indexing')
+                es.index(index=namespaced_index, body=indexing_record, id=index_start_str)
+                es.index(index=namespaced_index, body=indexing_record, id='latest_indexing')
             except Exception:
                 indexing_record['indexing_status'] = 'errored'
                 error_messages = copy.deepcopy(indexing_record['errors'])
                 del indexing_record['errors']
-                es.index(index=namespaced_index, doc_type='indexing', body=indexing_record, id=index_start_str)
-                es.index(index=namespaced_index, doc_type='indexing', body=indexing_record, id='latest_indexing')
+                es.index(index=namespaced_index, body=indexing_record, id=index_start_str)
+                es.index(index=namespaced_index, body=indexing_record, id='latest_indexing')
                 for item in error_messages:
                     if 'error_message' in item:
                         log.error('Indexing error', **item)
@@ -251,7 +251,17 @@ class Indexer(object):
         rev_linked_uuids = set()
         to_delete = []  # hold messages that will be deleted
         to_defer = []  # hold messages once we need to restart the worker
-        # max_sid does not change of the lifetime of request
+
+        # A note for the record of history: the idea behind this check is to store in
+        # ES the max_sid acquired at this time alongside the documents indexed into ES, so we have
+        # an "approximate" database time when the indexing occurred.
+        # What's going on here is when you make a DB edit that is successful, the sid assigned to that edit
+        # is passed to the indexing queue, and when that message is pulled down the indexer workers ask the
+        # DB what the current max_sid is, and if the msg sid is greater than the max_sid acquired at the start of
+        # the indexing run, it will re-drive the message to get processed by a worker who pulls down a "newer" sid.
+        # This is necessary because that indexer worker could have stale data in the embed cache. By having the message
+        # retrieved by a "newer" worker you are guaranteeing an up-to-date embed cache. A rare case indeed but still a
+        # possibility. - Will Sept 13 2022
         max_sid = request.registry[STORAGE].write.get_max_sid()
         deferred = False  # if true, we need to restart the worker
         messages, target_queue = self.get_messages_from_queue()
@@ -330,7 +340,7 @@ class Indexer(object):
                 # CHANGE - this needs to happen PER MESSAGE now
                 # add to secondary queue, if applicable
                 # search for all items that linkTo the non-strict items or contain
-                # a rev_link to to them
+                # a rev_link to them
                 if non_strict_uuids or rev_linked_uuids:
                     queued, failed = self.find_and_queue_secondary_items(non_strict_uuids,  # THIS IS NOW A SINGLE UUID
                                                                          rev_linked_uuids,
@@ -461,7 +471,7 @@ class Indexer(object):
             try:
                 namespaced_index = get_namespaced_index(request, result['item_type'])
                 self.es.index(
-                    index=namespaced_index, doc_type=result['item_type'], body=result,
+                    index=namespaced_index, body=result,
                     id=str(uuid), version=result['sid'], version_type='external_gte',
                     request_timeout=30
                 )