+++ snovault/tests/serverfixtures.py
@@ -129,6 +129,9 @@ def conn(engine_url):
     notice_pytest_fixtures(engine_url)
     engine_settings = {
         'sqlalchemy.url': engine_url,
+        'connect_args': {
+            'connection_timeout': 30
+        }
     }
 
     engine = configure_engine(engine_settings)
@@ -173,8 +176,11 @@ def external_tx(request, conn):
     notice_pytest_fixtures(request)
     # print('BEGIN external_tx')
     tx = conn.begin_nested()
-    yield tx
-    tx.rollback()
+    try:
+        yield tx
+        tx.rollback()
+    except Exception:
+        tx.rollback()  # conn does not implement .rollback()
     # # The database should be empty unless a data fixture was loaded
     # for table in Base.metadata.sorted_tables:
     #     assert conn.execute(table.count()).scalar() == 0
@@ -220,7 +226,7 @@ def zsa_savepoints(conn):
                 return
             if self.state == 'commit':
                 self.state = 'completion'
-                self.sp.commit()
+                # self.sp.commit() this call is not necessary and throwing exception in sqlalchemy
             else:
                 self.state = 'abort'
                 self.sp.rollback()